refactor(dx): Make repository self-explaining
