+++ docs/README.md
@@ -1,0 +1,8 @@
+# Documentation map
+
+This directory holds only knowledge the code cannot carry.
+
+- [`setup/`](./setup/) — runbooks for configuring external services such as Cloudflare, PostHog, and Tolgee.
+- [`recipes/`](./recipes/) — optional extension paths for template adopters. Durable rationale stays in prose; copyable implementation should be tested or pinned to a stable interface.
+
+Plans live in PR bodies, findings in issues, conventions in checks, and implementation facts in tests. See [`AGENTS.md`](./AGENTS.md) for the complete placement rule.
