--- docs/forms/change-password.json
@@ -1,66 +1,0 @@
-[
-	{
-		"id": "row-1",
-		"fields": [
-			{
-				"id": "field-1",
-				"name": "Password",
-				"type": "password",
-				"category": "password",
-				"label": "Current password",
-				"description": "",
-				"placeholder": "Enter current password",
-				"required": true,
-				"position": "full"
-			}
-		]
-	},
-	{
-		"id": "row-2",
-		"fields": [
-			{
-				"id": "field-2",
-				"name": "Password",
-				"type": "password",
-				"category": "password",
-				"label": "New password",
-				"description": "Minimum 10 characters with uppercase, lowercase, and number",
-				"placeholder": "Enter new password",
-				"required": true,
-				"position": "full"
-			}
-		]
-	},
-	{
-		"id": "row-3",
-		"fields": [
-			{
-				"id": "field-3",
-				"name": "Password",
-				"type": "password",
-				"category": "password",
-				"label": "Confirm new password",
-				"description": "",
-				"placeholder": "Confirm new password",
-				"required": true,
-				"position": "full"
-			}
-		]
-	},
-	{
-		"id": "row-4",
-		"fields": [
-			{
-				"id": "field-4",
-				"name": "Checkbox",
-				"type": "boolean",
-				"category": "checkbox",
-				"label": "Sign out of other devices",
-				"description": "For security, we recommend signing out of all other devices after changing your password",
-				"placeholder": "",
-				"required": false,
-				"position": "full"
-			}
-		]
-	}
-]
