--- docs/forms/forgot-password.json
@@ -1,18 +1,0 @@
-[
-	{
-		"id": "row-1",
-		"fields": [
-			{
-				"id": "field-1",
-				"name": "Email",
-				"type": "email",
-				"category": "email",
-				"label": "Email",
-				"description": "",
-				"placeholder": "m@example.com",
-				"required": true,
-				"position": "full"
-			}
-		]
-	}
-]
