--- docs/forms/login.json
@@ -1,34 +1,0 @@
-[
-	{
-		"id": "row-1",
-		"fields": [
-			{
-				"id": "field-1",
-				"name": "Email",
-				"type": "email",
-				"category": "email",
-				"label": "Email",
-				"description": "",
-				"placeholder": "m@example.com",
-				"required": true,
-				"position": "full"
-			}
-		]
-	},
-	{
-		"id": "row-2",
-		"fields": [
-			{
-				"id": "field-2",
-				"name": "Password",
-				"type": "password",
-				"category": "password",
-				"label": "Password",
-				"description": "",
-				"placeholder": "",
-				"required": true,
-				"position": "full"
-			}
-		]
-	}
-]
