--- docs/forms/reset-password.json
@@ -1,34 +1,0 @@
-[
-	{
-		"id": "row-1",
-		"fields": [
-			{
-				"id": "field-1",
-				"name": "Password",
-				"type": "password",
-				"category": "password",
-				"label": "New password",
-				"description": "Minimum 10 characters with uppercase, lowercase, and number",
-				"placeholder": "",
-				"required": true,
-				"position": "full"
-			}
-		]
-	},
-	{
-		"id": "row-2",
-		"fields": [
-			{
-				"id": "field-2",
-				"name": "Password",
-				"type": "password",
-				"category": "password",
-				"label": "Confirm new password",
-				"description": "",
-				"placeholder": "",
-				"required": true,
-				"position": "full"
-			}
-		]
-	}
-]
