+++ package.json
@@ -37,6 +37,7 @@
 		"i18n:pull": "varlock run -- bun run _tolgee:pull && prettier --write src/i18n/*.json",
 		"i18n:push": "varlock run -- bun run _tolgee:push",
 		"i18n:cleanup": "varlock run -- bun run _tolgee:cleanup",
+		"i18n:sync": "bun scripts/sync-locales.ts",
 		"_tolgee:pull": "TOLGEE_API_KEY=$VITE_TOLGEE_API_KEY tolgee pull",
 		"_tolgee:push": "TOLGEE_API_KEY=$VITE_TOLGEE_API_KEY tolgee push",
 		"_tolgee:cleanup": "TOLGEE_API_KEY=$VITE_TOLGEE_API_KEY tolgee tag --filter-not-extracted --filter-tag production --tag deprecated --untag production",
