+++ scripts/sync-locales.ts
@@ -1,0 +1,65 @@
+import { readFile, readdir, writeFile } from 'node:fs/promises';
+import path from 'node:path';
+import { fileURLToPath } from 'node:url';
+import { DEFAULT_LANGUAGE, SUPPORTED_LANGUAGES } from '../src/lib/i18n/languages';
+
+const root = path.resolve(path.dirname(fileURLToPath(import.meta.url)), '..');
+const translationsPath = path.join(root, 'src/lib/i18n/static-translations.generated.ts');
+const wranglerPath = path.join(root, 'wrangler.toml');
+
+export const localeCodes = SUPPORTED_LANGUAGES.map((language) => language.code);
+
+export function renderStaticTranslations(): string {
+	const imports = localeCodes
+		.map((code) => `import ${code} from '../../i18n/${code}.json';`)
+		.join('\n');
+	const entries = localeCodes.map((code) => `\t${code}`).join(',\n');
+
+	return `// Generated by \`bun run i18n:sync\`. Edit languages.ts and locale JSON files instead.\n${imports}\nimport type { LanguageCode } from './languages';\n\nexport const STATIC_TRANSLATIONS = {\n${entries}\n} as const satisfies Record<LanguageCode, Record<string, unknown>>;\n\nexport type SupportedLocale = keyof typeof STATIC_TRANSLATIONS;\nexport const SUPPORTED_LOCALES = Object.keys(STATIC_TRANSLATIONS) as SupportedLocale[];\n`;
+}
+
+export function workerRoutesLine(): string {
+	const routes = localeCodes.flatMap((code) => [`/${code}`, `/${code}/*`]);
+	return `run_worker_first = [${routes.map((route) => JSON.stringify(route)).join(', ')}]`;
+}
+
+export function renderWranglerConfig(current: string): string {
+	const next = current.replace(/^run_worker_first\s*=.*$/m, workerRoutesLine());
+	if (next === current && !current.includes(workerRoutesLine())) {
+		throw new Error('wrangler.toml has no run_worker_first assignment');
+	}
+	return next;
+}
+
+export async function assertLocaleFiles(): Promise<void> {
+	const localeDirectory = path.join(root, 'src/i18n');
+	const files = (await readdir(localeDirectory))
+		.filter((file) => file.endsWith('.json'))
+		.map((file) => path.basename(file, '.json'))
+		.sort();
+	const expected = [...localeCodes].sort();
+
+	if (JSON.stringify(files) !== JSON.stringify(expected)) {
+		throw new Error(
+			`Locale files do not match SUPPORTED_LANGUAGES. Expected ${expected.join(', ')}, found ${files.join(', ')}`
+		);
+	}
+	if (!localeCodes.includes(DEFAULT_LANGUAGE)) {
+		throw new Error(`DEFAULT_LANGUAGE ${DEFAULT_LANGUAGE} is not registered`);
+	}
+	if (new Set(localeCodes).size !== localeCodes.length) {
+		throw new Error('SUPPORTED_LANGUAGES contains duplicate codes');
+	}
+}
+
+async function sync(): Promise<void> {
+	await assertLocaleFiles();
+	await writeFile(translationsPath, renderStaticTranslations());
+	const wrangler = await readFile(wranglerPath, 'utf8');
+	await writeFile(wranglerPath, renderWranglerConfig(wrangler));
+	console.log(`Synced locales: ${localeCodes.join(', ')}`);
+}
+
+if (import.meta.main) {
+	await sync();
+}
