+++ src/lib/i18n/static-translations.generated.ts
@@ -1,0 +1,16 @@
+// Generated by `bun run i18n:sync`. Edit languages.ts and locale JSON files instead.
+import en from '../../i18n/en.json';
+import de from '../../i18n/de.json';
+import es from '../../i18n/es.json';
+import fr from '../../i18n/fr.json';
+import type { LanguageCode } from './languages';
+
+export const STATIC_TRANSLATIONS = {
+	en,
+	de,
+	es,
+	fr
+} as const satisfies Record<LanguageCode, Record<string, unknown>>;
+
+export type SupportedLocale = keyof typeof STATIC_TRANSLATIONS;
+export const SUPPORTED_LOCALES = Object.keys(STATIC_TRANSLATIONS) as SupportedLocale[];
